+++ R/rasterplot.R
@@ -67,33 +67,33 @@
 #' This allows to get same "user" to "pixel" coordinates conversion as the one used to create 'bg_'.
 #' @param ... other arguments to pass to graphics::plot().
 #' For example, providing xlim and/or ylim will controls if point will be shown or not
-#' @details some examples:
-#' set.seed(2)
-#' n_points = 1e7; n_clusters = 5
-#' x = c(t(sapply(1:5, FUN = function(i) rnorm(n_points / n_clusters, mean = sample(-2:2, size = 1), sd = 1/sample(1:10, 1)))))
-#' y = c(t(sapply(1:5, FUN = function(i) rnorm(n_points / n_clusters, mean = sample(-2:2, size = 1), sd = 1/sample(1:10, 1)))))
-#' # plot points
-#' rasterplot(x = x, y = y, col = "black")
-#' # generate img
-#' rasterplot(x = x, y = y, col = "black", draw = FALSE)
-#' # plot multiple shapes
-#' rasterplot(x = x, y = y, pch = c(3,5,9,10,2))
-#' # plot multiple shapes + colors
-#' bg_ = rasterplot(x = x, y = y, pch = c(3,5,9,10,2), col = c("plum", "green", "indianred", "blue", "black"))
-#' # addition of new points to an already drawn background, it a kind of points(...)
-#' rasterplot(x = x[1:1e5], y = y[1:1e5], col = "black", bg_ = bg_, bg_map = TRUE)
-#' # plot 1 shape  + multiple colors
-#' rasterplot(x = x, y = y, pch = ".", col = c("plum", "green", "indianred", "blue", "black"), force = TRUE)
-#' # density
-#' rasterplot(x = x, y = y, pch = 20, size = 7, draw = TRUE, col = colorRampPalette(c("blue", "green", "red"))(100))
-#' # density with limits
-#' rasterplot(x = x, y = y, draw = TRUE, xlim = c(0, 1.5), pntsonedge = FALSE, col = colorRampPalette(c("blue", "green", "red"))(100))
-#' # density with limits + computation on drawn points only
-#' rasterplot(x = x, y = y, draw = TRUE, xlim = c(0, 1.5), pntsonedge = TRUE, col = colorRampPalette(c("blue", "green", "red"))(100))
-#' # using rgba
-#' col = c("plum", "green", "indianred", "blue", "black")
-#' rgba = col2rgb(col, alpha = TRUE)
-#' rgba = t(apply(rgba, 1, FUN = function(x) rep(x, length.out = n_points)))
+#' @details some examples:\cr
+#' set.seed(2)\cr
+#' n_points = 1e7; n_clusters = 5\cr
+#' x = c(t(sapply(1:5, FUN = function(i) rnorm(n_points / n_clusters, mean = sample(-2:2, size = 1), sd = 1/sample(1:10, 1)))))\cr
+#' y = c(t(sapply(1:5, FUN = function(i) rnorm(n_points / n_clusters, mean = sample(-2:2, size = 1), sd = 1/sample(1:10, 1)))))\cr
+#' # plot points\cr
+#' rasterplot(x = x, y = y, col = "black")\cr
+#' # generate img\cr
+#' rasterplot(x = x, y = y, col = "black", draw = FALSE)\cr
+#' # plot multiple shapes\cr
+#' rasterplot(x = x, y = y, pch = c(3,5,9,10,2))\cr
+#' # plot multiple shapes + colors\cr
+#' bg_ = rasterplot(x = x, y = y, pch = c(3,5,9,10,2), col = c("plum", "green", "indianred", "blue", "black"))\cr
+#' # addition of new points to an already drawn background, it a kind of points(...)\cr
+#' rasterplot(x = x[1:1e5], y = y[1:1e5], col = "black", bg_ = bg_, bg_map = TRUE)\cr
+#' # plot 1 shape  + multiple colors\cr
+#' rasterplot(x = x, y = y, pch = ".", col = c("plum", "green", "indianred", "blue", "black"), force = TRUE)\cr
+#' # density\cr
+#' rasterplot(x = x, y = y, pch = 20, size = 7, draw = TRUE, col = colorRampPalette(c("blue", "green", "red"))(100))\cr
+#' # density with limits\cr
+#' rasterplot(x = x, y = y, draw = TRUE, xlim = c(0, 1.5), pntsonedge = FALSE, col = colorRampPalette(c("blue", "green", "red"))(100))\cr
+#' # density with limits + computation on drawn points only\cr
+#' rasterplot(x = x, y = y, draw = TRUE, xlim = c(0, 1.5), pntsonedge = TRUE, col = colorRampPalette(c("blue", "green", "red"))(100))\cr
+#' # using rgba\cr
+#' col = c("plum", "green", "indianred", "blue", "black")\cr
+#' rgba = col2rgb(col, alpha = TRUE)\cr
+#' rgba = t(apply(rgba, 1, FUN = function(x) rep(x, length.out = n_points)))\cr
 #' rasterplot(x = x, y = y, pch = ".", rgba = rgba, draw = TRUE)
 #' @return an [0, 255] integer array of (height, width, 4) of class `rasterplot`
 #' @keywords internal
@@ -224,17 +224,23 @@ rasterplot = function(x, y = NULL,
     img = cpp_raster(width = width, height = height, data, bg_)
     
     # subset img to drawing region
-    usr = par("usr")
-    lims = round(c(graphics::grconvertX(usr[1], "user", "ndc") * width,
-                   graphics::grconvertX(usr[2], "user", "ndc") * width,
-                   (1 - graphics::grconvertY(usr[3], "user", "ndc")) * height,
-                   (1 - graphics::grconvertY(usr[4], "user", "ndc")) * height))
+    usr = unlist(recursive = FALSE, use.names = FALSE, coordmap$domain)
+    lims = round(c(coord_to_px(coord = data.frame(x = usr[1:2], y = usr[3:4]),
+                               coordmap = coordmap,
+                               pntsonedge = F)) + c(1,1,0,0))
     
     # add image to plot background
-    rasterImage(image = cpp_as_nativeRaster(img[lims[3]:lims[4], lims[1]:lims[2],]),
+    if(!identical(get_coordmap_adjusted(), coordmap)) {
+      text(x = graphics::grconvertX(0.5, "npc", "user"),
+           y = graphics::grconvertY(0.5, "npc", "user"),
+           labels = "you should not modify graphic device while drawing raster",
+           col = "red")
+      img = NULL
+    } else {
+      rasterImage(image = cpp_as_nativeRaster(img[lims[3]:lims[4], lims[1]:lims[2],]),
                 xleft = usr[1], xright = usr[2], ybottom = usr[4], ytop = usr[3],
                 interpolate = interpolate)
-    
+    }
     return(invisible(structure(img, "coordmap" = coordmap, class = "rasterplot")))
   } else {
     return(invisible(structure(cpp_raster(width = width, height = height, data, bg_), "coordmap" = coordmap, class = "rasterplot")))